feat(demo): refacture CarmenView and add useCarmenView hook , add favicon
